mozilla/captain#25: Add heartbeat and restructure listeners.
Because we want to send the heartbeat message and listen for commands
in a single process, we have to switch to using threads for shove. And
if we’re using threads, we don’t need the asynchronous RabbitMQ adapters
anymore, which helps simplify the threaded code.

On launch, shove now launches two threads and waits for signals in the
main thread. One of the threads sends heartbeats back to Captain on a
regular basis, while the other listens for and executes commands.